SCRS-5518 Moved compliance introduction to new logical group packaging
