+++ app/views/pages/compliance_introduction.scala.html
@@ -16,7 +16,7 @@
 
         <br />
 
-        @form(action = controllers.userJourney.routes.ComplianceIntroductionController.submit()) {
+        @form(action = controllers.userJourney.sicAndCompliance.routes.ComplianceIntroductionController.submit()) {
             <div class="form-group">
                 <button class="button-get-started" type="submit" role="button" id="continue">@messages("app.common.continue")</button>
             </div>
